+++ CHANGES.rst
@@ -1,6 +1,9 @@
 0.8 (Unreleased)
 ----------------
 
+- ``LabelMapperRange`` now returns ``LabelMapperRange._no_label`` when the key is
+  not within any range. [#71]
+
 0.7 (2016-12-23)
 ----------------